Specify Epoll as External (#28)
As specified in the vite-plugin-electron docs, we might need to manually specify certain native C/C++ modules because rollup will not build them properly. This is now done for epoll, which is the native library used by gpio to control the Pi GPIO. This is important as it makes the Garage buttons work.